+++ search.php
@@ -16,6 +16,8 @@
 			
 		<?php endwhile; ?>
 		
+		<?php get_template_part( 'nav', 'archive' ); ?>
+		
 	<?php endif; ?>
 	
 </main>